feat : 대시 컴포넌트 함수 이동
