+++ client/src/component/Board/Dash.jsx
@@ -1,5 +1,6 @@
 //모듈
 import styled from 'styled-components';
+import { useRef, useEffect, useState } from 'react';
 
 //공통 스타일
 import { COLOR } from '../../style/theme';
@@ -102,11 +103,60 @@ const CommentIcon = styled(BiComment)`
   color: ${COLOR.main_dark_blue};
 `;
 
-const Dash = ({ posts }) => {
+const EndDetect = styled.div`
+  width: 100%;
+  height: 20px;
+`;
+
+const Dash = ({ posts, setCurrentPage, isDash }) => {
+  const [data, setData] = useState([]);
+  const observer = useRef(null);
+  const sentinelRef = useRef(null);
+
+  useEffect(() => {
+    const options = {
+      root: null,
+      rootMargin: '0px',
+      threshold: 0.5, // 타겟의 50%가 보이는 시점에서 감지
+    };
+
+    const handleObserver = (entries) => {
+      const target = entries[0];
+      if (target.isIntersecting) {
+        handleLoadMore();
+      }
+    };
+
+    observer.current = new IntersectionObserver(handleObserver, options);
+
+    // 컴포넌트가 마운트될 때 신호 타겟을 관찰
+    if (observer.current) {
+      observer.current.observe(sentinelRef.current);
+    }
+
+    // 컴포넌트가 언마운트될 때 옵저버를 정리
+    return () => {
+      if (observer.current) {
+        observer.current.disconnect();
+      }
+    };
+  }, []);
+
+  useEffect(() => {
+    // posts 상태가 업데이트될 때마다 `data`에 추가
+    setData((prevData) => [...prevData, ...posts]);
+  }, [posts]);
+
+  const handleLoadMore = () => {
+    if (isDash) {
+      setCurrentPage((prev) => prev + 1);
+    }
+  };
+
   return (
     <Container>
-      {posts.length
-        ? posts.map((post) => (
+      {data.length
+        ? data.map((post) => (
             <DashBorad key={post.id}>
               <Image>
                 <img src={post.image} alt="캘린더 이미지" />
@@ -127,6 +177,7 @@ const Dash = ({ posts }) => {
             </DashBorad>
           ))
         : null}
+      <EndDetect ref={sentinelRef} />
     </Container>
   );
 };
